update:: 재시작 시 엔진, 월드, 점수 초기화 #98
