+++ components/Page/SuikaGame/HUD/index.tsx
@@ -9,12 +9,10 @@ type HUDProps = {
 
 export default function HUD({ score, nextItem }: HUDProps) {
   return (
-    <div className="max-w-[270px] w-[100%] h-[75px] flex flex-col rounded border-4 border-[#6F4927]">
-      {/* <div className="w-full h-1/3 bg-[#FFDCC8] text-[#6F4927]"> */}
+    <div className="w-[94%] md:max-w-[270px] md:w-[100%] h-[90px] md:h-[75px] flex flex-col rounded border-4 border-[#6F4927]">
       <p className="w-full h-1/3 bg-[#FFDCC8] text-[#6F4927] text-end text-[1.2rem] font-extrabold pr-[10px]">
         Next
       </p>
-      {/* </div> */}
       <div className="w-full h-2/3 flex justify-between items-center bg-[#6F4927] pl-[2px] pr-[10px]">
         <p className="text-[2rem] text-[#FFDCC8] font-black pl-[10px]">{score}</p>
         <Image
@@ -27,7 +25,6 @@ export default function HUD({ score, nextItem }: HUDProps) {
           draggable="false"
           className="pt-2"
         />
-        {/* <div className="flex flex-col justify-center items-center gap-4"> */}
       </div>
     </div>
   )
